requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosmtplib==2.0.2
alembic==1.13.1
amqp==5.2.0
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==42.0.2
dnspython==2.5.0
easyschedule==0.107
email-validator==2.1.0.post1
fastapi==0.109.0
fastapi-cache2==0.2.1
fastapi-mail==1.4.1
fastapi-users==12.1.3
fastapi-users-db-sqlalchemy==6.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
humanize==4.9.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
kombu==5.3.7
makefun==1.15.2
Mako==1.3.2
MarkupSafe==2.1.4
multidict==6.0.5
orjson==3.9.12
passlib==1.7.4
pendulum==3.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
pydantic==2.5.3
pydantic-extra-types==2.4.1
pydantic-settings==2.1.0
pydantic_core==2.14.6
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.7
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
schedule==1.2.1
setuptools==69.5.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.35.1
time-machine==2.14.1
tornado==6.4
typing_extensions==4.9.0
tzdata==2024.1
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.0
vine==5.1.0
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
yarl==1.9.4
